requirements.txt
asfpy~=0.32
ezt~=1.1
Flask~=2.0.2
python-dotenv~=0.19.2
click~=8.0.3
pytest~=7.0.0
pytest-dotenv~=0.5.2
pytest-datadir~=1.3.1
pytest-flask~=1.2.0
coverage~=6.4.1
pytest-cov~=3.0.0
pyyaml~=6.0